requirements.txt
boto3==1.9.242
botocore==1.12.242
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
idna==2.8
jmespath==0.9.4
mysql-connector==2.2.9
python-dateutil==2.8.0
requests==2.22.0
rollbar==0.14.7
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6